Factor out `edit`
