+++ async-openai/src/types/mod.rs
@@ -5,6 +5,7 @@ mod assistant_file;
 mod assistant_impls;
 mod audio;
 mod common;
+mod edit;
 mod file;
 mod fine_tune;
 mod image;
@@ -20,6 +21,7 @@ pub use assistant::*;
 pub use assistant_file::*;
 pub use audio::*;
 pub use common::*;
+pub use edit::*;
 pub use file::*;
 pub use fine_tune::*;
 pub use image::*;
